requirements.txt
appdirs==1.4.3
Django==1.11
django-betterforms==1.1.4
django-bootstrap3==8.2.2
django-bootstrap3-datepicker==0.4.1
django-bootstrap3-datetimepicker==2.2.3
django-datetime-widget==0.9.3
django-extensions==1.7.8
django-forms-bootstrap==3.0.1
django-timezone-field==1.2
djangorestframework==3.6.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
